afxdp/flow_cache: relocate inline tests to flow_cache_tests.rs
flow_cache.rs is a flat file. 923-LOC inline test block extracted to sibling flow_cache_tests.rs (926 LOC incl 5-line header). Stub: `#[cfg(test)] #[path = "flow_cache_tests.rs"] mod tests;`.

After: flow_cache.rs drops from 1,370 → 450 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
